Add rack_env production, add seed task, update seeds to not execute when we have already most current data.

+++ db/seeds.rb
@@ -3,8 +3,9 @@ EVENTS = %w(TAB_OPEN TAB_CLOSE TAB_RELOAD)
 DESCS = ['New tab was opened.', 'Tab was closed.', 'Tab was reloaded.']
 
 # Clear events table
-Event.delete_all
-
-EVENTS.each_with_index do |event, index|
-  Event.create(name: event, desc: DESCS[index])
+if Event.count != EVENTS.size
+  Event.delete_all
+  EVENTS.each_with_index do |event, index|
+    Event.create(name: event, desc: DESCS[index])
+  end
 end
